requirement.txt
pytorch
transformers
pytorch_pretrained_bert
scipy
